Move clean_html() & foreground_color() to utilities.html
