+++ netbox/utilities/templatetags/builtins/filters.py
@@ -11,9 +11,9 @@ from markdown import markdown
 from markdown.extensions.tables import TableExtension
 
 from netbox.config import get_config
+from utilities.html import clean_html, foreground_color
 from utilities.markdown import StrikethroughExtension
 from utilities.string import title
-from utilities.utils import clean_html, foreground_color
 
 __all__ = (
     'bettertitle',
